requirements.txt
-f /usr/share/pip-wheels
asgiref==3.7.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.25
django-cors-headers==4.3.1
django-import-export==2.7.0
django-ipware==6.0.4
django-rest-authemail==2.1.7
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.6
MarkupPy==1.14
mysqlclient==2.2.4
odfpy==1.4.1
openpyxl==3.1.2
pillow==11.0.0
python-decouple==3.8
python-ipware==2.0.2
pytz==2024.1
PyYAML==6.0.1
reportlab==4.3.1
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
typing-extensions==4.10.0
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
